even more cleanup
