--- Assets/ScriptableObjects/Signals/DialogueChannel.asset
@@ -1,14 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 142d4f5760b924e50b0f6c35ff81616a, type: 3}
-  m_Name: DialogueChannel
-  m_EditorClassIdentifier:
