--- Assets/ScriptableObjects/Signals/DialogueChannel.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 804f7b18096194b01997ce8366ce4628
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
