--- Assets/ScriptableObjects/Signals/PlayerDeatchEventChannel.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 5f6568f6334dc4636823e7fe78c905f2
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
