+++ Assets/Scripts/Spells/Ghost.cs.meta
