+++ Assets/Scripts/Spells/Health.cs
@@ -2,7 +2,7 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-[CreateAssetMenu]
+[CreateAssetMenu(menuName = "Spells/Health")]
 public class Health : Spell
 {
     private int OriginalHealth;
